+++ rust/writers/to_tiles/mod.rs
@@ -252,7 +252,7 @@ impl<W: TileWriter> TileBuilder<W> {
             } else {
                 self.tile_writer.write_tile_wm(zoom, x, y, data);
                 let (w, s, e, n) =
-                    xyz_to_bbox(x, y, zoom, Some(false), Some(wm::Source::WGS84), None);
+                    xyz_to_bbox(x, y, zoom as f64, Some(false), Some(wm::Source::WGS84), None);
                 self.meta_builder.add_tile_wm(zoom, x, y, &LonLatBounds::new(w, s, e, n));
             }
         }
